tech-debt(arch): extend the 2000-LOC cap to tsz-binder and tsz-cli (#16733)
CLAUDE.md documents the 2000-LOC cap repo-wide, but `LINE_LIMIT_CHECKS`
in scripts/arch/arch_guard_shared.py only ever registered
`tsz-checker/src`, leaving every other crate's drift invisible to
`arch-size`. Per #16733's own "nearly free wins" note, split the one
tsz-binder file and two tsz-cli files that were barely over the cap,
then registered both crates:

- crates/tsz-binder/src/state/core.rs (2006 lines) split: the lib-merge
  convenience wrappers and incremental-rebind entry point
  (merge_lib_symbols, bind_source_file_with_libs,
  bind_source_file_incremental, prune_incremental_maps,
  recompute_module_export_equals_non_module) moved to a new
  state/core_incremental.rs.
- crates/tsz-cli/src/bin/tsz_server/main.rs (2038 lines) split: the
  byte-level bracket/regex-literal scanning helpers used by structure
  navigation (build_code_map, scan_forward, scan_backward,
  find_angle_bracket_match, is_regex_literal_start) moved to a new
  bin/tsz_server/bracket_match.rs, matching the existing handlers_*
  split convention for that binary.
- crates/tsz-cli/src/driver/check_utils/tests.rs (2258 lines) split at
  a clean test boundary into tests.rs + tests_part2.rs (child module,
  same convention as the crate's other `_part2` test splits).

Also pinned coverage so it cannot silently regress:
scripts/arch/test_arch_guard.py gains
test_line_limit_checks_cover_at_least_these_crates, asserting
tsz-checker, tsz-binder, and tsz-cli stay registered in
LINE_LIMIT_CHECKS.

Scope: this covers 2 of the 6 unguarded crates named in #16733 (11
files in tsz-solver and 10 in tsz-emitter remain, plus 3 in tsz-core);
sized as one slice of the crate-by-crate campaign the issue suggests,
not the full campaign.
